third_party/blink/web_tests/external/wpt/editing/run/justifycenter_1-1000-expected.txt

This is a testharness.js-based test.
Found 126 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foobar</div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foobar</div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[]bar<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<div style=\\"text-align:center\\">foobar</div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<div style=\\"text-align:center\\">foobar</div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<div style=\\"text-align:center\\">foobar</div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<div style=\\"text-align:center\\">foobar</div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[bar]baz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[bar]baz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[bar]baz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[bar]baz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foobarbaz</div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar]baz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar]baz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foobarbaz</div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar]baz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar]baz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foobar<b>bazqoz</b>quz</div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar<b>bazqoz</b>quz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foobar<b>bazqoz</b>quz</div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar<b>bazqoz</b>quz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foobar</p></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foobar</p></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foobar</p></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foobar</p></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foobarbaz</p></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foobarbaz</p></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foobarbaz</p></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foobarbaz</p></div><p>extra</p>" but got "<p style=\\"text-align:center\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><h1>foobarbaz</h1></div><p>extra</p>" but got "<h1 style=\\"text-align:center\\">foobarbaz</h1><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><h1>foobarbaz</h1></div><p>extra</p>" but got "<h1 style=\\"text-align:center\\">foobarbaz</h1><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><h1>foobarbaz</h1></div><p>extra</p>" but got "<h1 style=\\"text-align:center\\">foobarbaz</h1><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><h1>foobarbaz</h1></div><p>extra</p>" but got "<h1 style=\\"text-align:center\\">foobarbaz</h1><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><pre>foobarbaz</pre></div><p>extra</p>" but got "<pre style=\\"text-align:center\\">foobarbaz</pre><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><pre>foobarbaz</pre></div><p>extra</p>" but got "<pre style=\\"text-align:center\\">foobarbaz</pre><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><pre>foobarbaz</pre></div><p>extra</p>" but got "<pre style=\\"text-align:center\\">foobarbaz</pre><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><pre>foobarbaz</pre></div><p>extra</p>" but got "<pre style=\\"text-align:center\\">foobarbaz</pre><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><xmp>foobarbaz</xmp></div><p>extra</p>" but got "<xmp style=\\"text-align:center\\">foobarbaz</xmp><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><xmp>foobarbaz</xmp></div><p>extra</p>" but got "<xmp style=\\"text-align:center\\">foobarbaz</xmp><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><xmp>foobarbaz</xmp></div><p>extra</p>" but got "<xmp style=\\"text-align:center\\">foobarbaz</xmp><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><xmp>foobarbaz</xmp></div><p>extra</p>" but got "<xmp style=\\"text-align:center\\">foobarbaz</xmp><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifycenter",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["justifycenter",""]] "<center><p>[foo]<p>bar</center><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["justifycenter",""]] "<center><p>[foo]<p>bar</center><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foo</p><p>bar</p></div><p>extra</p>" but got "<center><p>foo</p><p>bar</p></center><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foo</p><p>bar</p></div><p>extra</p>" but got "<center><p>foo</p><p>bar</p></center><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\"><p>foo</p><p>bar</p></div><p>extra</p>" but got "<center><p>foo</p><p>bar</p></center><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifycenter",""]] "<center><p>[foo<p>bar]</center><p>extra" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
Harness: the test ran to completion.
